fix(ios): patch RNFB Firestore objc imports for Xcode 26

+++ gps-tracker-app/plugins/withRNFBFirestoreHeaderFix.js
@@ -3,7 +3,8 @@ const fs = require('fs')
 const path = require('path')
 
 // Xcode 26 + modular headers can require RNFBAppModule to be imported before RCTBridgeModule
-// in RNFB Firestore headers. This patch is applied automatically during prebuild.
+// in RNFB Firestore headers. Also, some .m files need an explicit RCTBridgeModule import
+// so RCT_EXPORT_* macros resolve correctly.
 module.exports = function withRNFBFirestoreHeaderFix(config) {
   return withDangerousMod(config, [
     'ios',
@@ -22,13 +23,13 @@ module.exports = function withRNFBFirestoreHeaderFix(config) {
         return config
       }
 
-      const files = fs
+      const headerFiles = fs
         .readdirSync(headersDir)
         .filter((f) => f.endsWith('.h'))
         .map((f) => path.join(headersDir, f))
 
       let changed = 0
-      for (const file of files) {
+      for (const file of headerFiles) {
         const src = fs.readFileSync(file, 'utf8')
         if (!src.includes('#import <React/RCTBridgeModule.h>')) continue
 
@@ -38,6 +39,35 @@ module.exports = function withRNFBFirestoreHeaderFix(config) {
           '#import <React/RCTBridgeModule.h>',
           '#import <RNFBApp/RNFBAppModule.h>\n#import <React/RCTBridgeModule.h>'
         )
+
+        if (next !== src) {
+          fs.writeFileSync(file, next, 'utf8')
+          changed += 1
+        }
+      }
+
+      const objcFiles = [
+        'RNFBFirestoreModule.m',
+        'RNFBFirestoreTransactionModule.m',
+        'RNFBFirestoreDocumentModule.m',
+        'RNFBFirestoreCollectionModule.m',
+      ].map((f) => path.join(headersDir, f))
+
+      for (const file of objcFiles) {
+        if (!fs.existsSync(file)) continue
+        const src = fs.readFileSync(file, 'utf8')
+        if (src.includes('#import <React/RCTBridgeModule.h>')) continue
+
+        const marker = '#import <React/RCTUtils.h>'
+        let next = src
+        if (src.includes(marker)) {
+          next = src.replace(
+            marker,
+            `${marker}\n#import <React/RCTBridgeModule.h>`
+          )
+        } else {
+          next = `#import <React/RCTBridgeModule.h>\n${src}`
+        }
 
         if (next !== src) {
           fs.writeFileSync(file, next, 'utf8')
@@ -55,4 +85,3 @@ module.exports = function withRNFBFirestoreHeaderFix(config) {
     },
   ])
 }
-
